+++ src/device/sysdep_NETBSD.c
@@ -83,7 +83,7 @@
 
 
 static struct {
-        uint64_t timestamp;
+        unsigned long long timestamp;
         size_t diskCount;
         size_t diskLength;
         struct io_sysctl *disk;
@@ -114,7 +114,7 @@ static bool _parseDevice(const char *path, Device_T device) {
 }
 
 
-static bool _getStatistics(uint64_t now) {
+static bool _getStatistics(unsigned long long now) {
         // Refresh only if the statistics are older then 1 second (handle also backward time jumps)
         if (now > _statistics.timestamp + 1000 || now < _statistics.timestamp - 1000) {
                 size_t len = 0;
@@ -145,7 +145,7 @@ static bool _getDummyDiskActivity(void *_inf) {
 
 static bool _getBlockDiskActivity(void *_inf) {
         Info_T inf = _inf;
-        uint64_t now = Time_milli();
+        unsigned long long now = Time_milli();
         bool rv = _getStatistics(now);
         if (rv) {
                 for (int i = 0; i < _statistics.diskCount; i++)     {
@@ -190,9 +190,9 @@ static bool _compareDevice(const char *device, struct statvfs *mnt) {
 }
 
 
-static void _filesystemFlagsToString(Info_T inf, uint64_t flags) {
+static void _filesystemFlagsToString(Info_T inf, unsigned long long flags) {
         struct mystable {
-                uint64_t flag;
+                unsigned long long flag;
                 char *description;
         } t[]= {
 #ifdef MNT_DISCARD
